split routes for eas into a separate file. moved reading validation to its own file in extensions
