+++ extensions/__init__.py
@@ -5,6 +5,7 @@ Import data from files to initiate and more easily pass through
 
 from extensions.arcplus import ArcPlus
 from extensions.flask_addons import db, login_manager, ma
+from extensions.validate_readings import get_valid_readings
 # from extensions.config_setup import load_yaml, update_yaml, yaml_file_path
 # from extensions.custom_exceptions import CustomException
 # from extensions.env_path import env_path
