+++ instance/.env
@@ -1,3 +1,5 @@
 FLASK_SQLALCHEMY_DATABASE_URI='sqlite:///readings.db'
 FLASK_SQLALCHEMY_TRACK_MODIFICATIONS=false
 FLASK_DEBUG=true      # Only for use in development
+FLASK_SECRET_KEY='8f086390cfbf85f2ff9bde1687502b2a'
+FLASK_ENCRYPTION_KEY='OTCh-NDbHxdpQmlNDkKn4fdE1q-_h8lfQ98NuZBXm3g='
